Fixes #46 (published state for records)
Setting record state (cache.recstate) to published in the sqlite3 database
required looking at the progress sqlite3 databases generated by the App
Engine bulkloader tool for both the Record and RecordIndex. These databases
have a state column that equals 2 if the row in the CSV file was bulkloaded
successfully.

The magic happens in bulkload.py where a generator steps through each
line in report.csv, grabs the corresponding rows from the progress databases,
then yields the (reckey, state) 2-tuples. State is set to published but is
changed to error if state column in the progress database rows are not both
equal to 2. The generator is executed with a cache update query to update
recstate for the given reckey.